# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: audit — API fiscal (ESAA-Flow)
  version: 1.0.0
  description: |
    Contrato mínimo para destravar o frontend (Lovable). Toda escrita passa pelo
    orquestrador single-writer: o cliente envia uma *intenção*; a API valida
    (7 camadas), grava o evento e re-projeta. Leituras vêm da projeção.

    Convenções:
    - Identificadores em inglês; mensagens de erro em PT-BR.
    - `cnpj` sempre com 14 dígitos, sem máscara.
    - `period` no formato `YYYY-MM` (competência).
    - Toda resposta de escrita devolve o `event_seq` gravado e o `projection_hash`.
servers:
  - url: http://localhost:3000/v1
    description: desenvolvimento
  - url: https://api.SEU-DOMINIO.com.br/v1
    description: produção
security:
  - bearerAuth: []
...
paths:
...
  /clients/{cnpj}/certificate:
    parameters:
      - $ref: "#/components/parameters/Cnpj"
    get:
      tags: [certificates]
      summary: Metadados do certificado A1 do cliente (nunca retorna o PFX)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CertificateInfo" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [certificates]
      summary: Armazena ou substitui o certificado (emite `certificate.stored`; PFX cifrado em repouso)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [pfx, password]
              properties:
                pfx: { type: string, format: binary }
                password: { type: string, format: password }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WriteResult" }
        "422": { $ref: "#/components/responses/Rejected" }
    delete:
      tags: [certificates]
      summary: Remove o certificado (emite `certificate.removed`)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WriteResult" }
  /clients/{cnpj}/certificate/usage:
    parameters:
      - $ref: "#/components/parameters/Cnpj"
    get:
      tags: [certificates]
      summary: Log de uso do certificado (projeção dos eventos `certificate.used`)
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/CertificateUsage" }
                  total: { type: integer }
  /certificates/expiring:
    get:
      tags: [certificates]
      summary: Certificados da carteira que vencem nos próximos N dias
      parameters:
        - in: query
          name: days
          schema: { type: integer, default: 60, minimum: 1, maximum: 365 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CertificateInfo" }

  # ----------------------------------------------------------- assistant
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Cnpj:
      in: path
      name: cnpj
      required: true
      # Doze posições alfanuméricas e dois dígitos verificadores numéricos: a
      # Receita emite CNPJ alfanumérico desde 31/07/2026 (IN RFB 2.229/2024).
      # O dígito verificador não é conferido pelo schema — JSON Schema não faz
      # módulo 11 —, e sim na fronteira de entrada da API.
      schema: { type: string, pattern: "^[0-9A-Za-z]{12}[0-9]{2}$" }
...
    Page:
      in: query
      name: page
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      in: query
      name: page_size
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
  responses:
...
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Rejected:
      description: Intenção rejeitada pelo pipeline (evento `output.rejected` gravado)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Rejection" }
  schemas:
...
    PeriodState:
      type: string
      enum: [open, assessed, reconciled, confirmed]
...
    RejectionReason:
      type: string
      enum:
        - schema_violation
        - unknown_code
        - code_incompatible
        - regime_violation
        - closed_period_violation
        - fisco_mismatch
        - duplicate_document
        - sequence_gap
        - verification_mismatch
        - boundary_violation
...
    CertificateInfo:
      type: object
      properties:
        cnpj: { type: string }
        subject: { type: string }
        issuer: { type: string }
        serial: { type: string }
        valid_from: { type: string, format: date-time }
        valid_to: { type: string, format: date-time }
        days_to_expiry: { type: integer }
        stored_at: { type: string, format: date-time }
        stored_by: { type: string }
        last_used_at: { type: string, format: date-time, nullable: true }
        usage_count:
          type: integer
          description: |
            Usos desde que **este** certificado foi guardado. Zera quando um novo
            PFX substitui o anterior. O campo se chamava `usage_count_30d` e
            contava tudo — não havia janela de 30 dias em lugar nenhum.
    CertificateUsage:
      type: object
      description: Projeção de um evento `certificate.used`
      properties:
        event_seq: { type: integer }
        used_at: { type: string, format: date-time }
        actor:
          type: object
          properties:
            type: { type: string, enum: [user, agent, system] }
            id: { type: string }
        purpose:
          type: string
          enum: [dfe_distribution, nfse_query, manifestation, sped_transmission, other]
        target: { type: string, description: Serviço/URL acessado }
        outcome: { type: string, enum: [success, failure] }
        ip: { type: string, nullable: true }
...
    WriteResult:
      type: object
      properties:
        event_id: { type: string, format: uuid }
        event_seq: { type: integer }
        action: { type: string }
        projection_hash: { type: string }
        state: { $ref: "#/components/schemas/PeriodState" }
    Rejection:
      type: object
      properties:
        rejected: { type: boolean, enum: [true] }
        filename:
          type: string
          description: |
            Arquivo que foi recusado. Presente nas rejeições do `207` de
            ingestão — sem ele, num lote de 200 XMLs a tela sabe que houve
            recusa e não sabe de qual arquivo.
        layer: { type: integer, minimum: 1, maximum: 7 }
        reason: { $ref: "#/components/schemas/RejectionReason" }
        message: { type: string }
        details: { type: object }
        rejection_event_seq: { type: integer, description: Seq do `output.rejected` gravado }
    Error:
      type: object
      properties:
        code: { type: string }
        message: { type: string }